requirements.txt
blinker==1.7.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
Werkzeug==3.0.1
zipp==3.17.0
